+++ gridsome.server.js
@@ -20,12 +20,4 @@ module.exports = function (api) {
       options.videoUrl = options.videoUrl|| '';
     }
   })
-
-  api.afterBuild (({ redirects }) => {
-    for (const rule of redirects) {
-      rule.from = '/checkout/:orderId'
-      rule.to = '/checkout/_order_id.html'
-      rule.status = 200
-    }
-  })
 }
