+++ src/components/Partials/Footer.vue
@@ -2,9 +2,7 @@
   <div class="bg-gray-800 dark:bg-gray-900 text-white dark:text-gray-400 text-center text-sm">
     <div class="container mx-auto py-16">
       <div class="mb-4">
-        <p>
-         Copyright {{ new Date().getFullYear() }} by {{ $static.metadata.siteName}} &middot; Powered by
-          <a href="https://www.gridsome.org" target="_blank" >Gridsome</a> </p> 
+         Copyright {{ new Date().getFullYear() }} by {{ $static.metadata.siteName}}
       </div>
       <div class="mb-4">
         <ul class="list-reset flex justify-center">
